heavenWebapp/HeavenBackend-develop/requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
black==23.3.0
Brotli==1.0.9
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
common==0.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
cssselect2==0.7.0
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==3.2.10
django-allauth==0.54.0
django-apscheduler==0.5.0
django-cors-headers==4.2.0
django-mysql==4.9.0
django-rest-auth==0.9.5
django-reversion==5.0.4
django-simple-history==3.3.0
djangorestframework==3.12.0
djangorestframework-dataclasses==1.2.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
fonttools==4.41.0
fpdf==1.7.2
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
mariadb-dyncol==3.6.1
MarkupSafe==2.1.2
modules==1.0.0
msgpack==1.0.5
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
pdfkit==1.0.0
Pillow==10.0.0
platformdirs==3.8.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.7.0
PyJWT==1.7.1
pyOpenSSL==23.2.0
pyphen==0.14.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==5.0.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tabulate==0.9.0
tinycss2==1.2.1
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
weasyprint==59.0
webencodings==0.5.1
zope.interface==6.0
zopfli==0.2.2
